requirements.txt
# Core RAG and LLM dependencies
llama-index==0.10.12
llama-index-llms-openai==0.1.6
llama-index-embeddings-huggingface==0.1.4
openai==1.12.0
groq
httpx==0.27.2

# Knowledge Graph
neo4j==5.16.0

# Embeddings
sentence-transformers==2.3.1

# Data processing
pandas==2.2.0
numpy<2
beautifulsoup4==4.12.3
pypdf==4.0.1
python-docx==1.1.0
markdown==3.5.2
python-dotenv
pyvis

# Demo interface
streamlit==1.31.0
pyvis==0.3.2

# Evaluation
rouge-score==0.1.2
nltk==3.8.1

# Utilities
python-dotenv==1.0.1
tqdm==4.66.1
